Move bibtex alignment function to its own file
