+++ src/main.ts
@@ -29,6 +29,7 @@ import {DefinitionProvider} from './providers/definition'
 import {LatexFormatterProvider} from './providers/latexformatter'
 import {FoldingProvider} from './providers/folding'
 import { SnippetPanel } from './components/snippetpanel'
+import { BibtexFormater } from './components/bibtexformater'
 
 import {checkDeprecatedFeatures, newVersionMessage, obsoleteConfigCheck} from './config'
 
@@ -117,9 +118,9 @@ export function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand('latex-workshop.showCompilationPanel', () => extension.buildInfo.showPanel())
     vscode.commands.registerCommand('latex-workshop.showSnippetPanel', () => extension.snippetPanel.showPanel())
 
-    vscode.commands.registerCommand('latex-workshop.bibsort', () => extension.commander.bibtexFormat(true, false))
-    vscode.commands.registerCommand('latex-workshop.bibalign', () => extension.commander.bibtexFormat(false, true))
-    vscode.commands.registerCommand('latex-workshop.bibalignsort', () => extension.commander.bibtexFormat(true, true))
+    vscode.commands.registerCommand('latex-workshop.bibsort', () => extension.bibtexFormater.bibtexFormat(true, false))
+    vscode.commands.registerCommand('latex-workshop.bibalign', () => extension.bibtexFormater.bibtexFormat(false, true))
+    vscode.commands.registerCommand('latex-workshop.bibalignsort', () => extension.bibtexFormater.bibtexFormat(true, true))
 
     context.subscriptions.push(vscode.workspace.onDidSaveTextDocument( (e: vscode.TextDocument) => {
         if (extension.manager.hasTexId(e.languageId)) {
@@ -294,6 +295,7 @@ export class Extension {
     snippetPanel: SnippetPanel
     graphicsPreview: GraphicsPreview
     mathPreview: MathPreview
+    bibtexFormater: BibtexFormater
 
     constructor() {
         this.extensionRoot = path.resolve(`${__dirname}/../../`)
@@ -318,6 +320,7 @@ export class Extension {
         this.pegParser = new PEGParser(this)
         this.graphicsPreview = new GraphicsPreview(this)
         this.mathPreview = new MathPreview(this)
+        this.bibtexFormater = new BibtexFormater(this)
         this.logger.addLogMessage('LaTeX Workshop initialized.')
     }
 }
